# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/edgesum/page/network/DeviceDetailFragment.java
#	app/src/main/java/com/example/edgesum/page/network/WiFiDirectActivity.java
#	app/src/main/java/com/example/edgesum/util/network/WiFiDirectBroadcastReceiver.java
#	app/src/main/java/com/example/edgesum/util/video/viewholderprocessor/RawFootageViewHolderProcessor.java
#	app/src/main/res/layout/activity_device_detail_fragment.xml

+++ app/src/main/java/com/example/edgesum/util/video/videoeventhandler/SummarisedVideosEventHandler.java
@@ -23,19 +23,51 @@ public class SummarisedVideosEventHandler implements VideoEventHandler {
     @Subscribe
     @Override
     public void onAdd(AddEvent event) {
+        if (event == null) {
+            Log.e(TAG, "Null event");
+            return;
+        }
+        if (repository == null) {
+            Log.e(TAG, "Null repository");
+            return;
+        }
+        if (event.video == null) {
+            Log.e(TAG, "Null video");
+            return;
+        }
 
         if (event.type == Type.SUMMARISED) {
-            Log.i(TAG, "onAdd");
-            repository.insert(event.video);
+            Log.v(TAG, "onAdd");
+            try {
+                repository.insert(event.video);
+            } catch (Exception e) {
+                Log.e(TAG, String.format("onAdd error: \n%s", e.getMessage()));
+            }
         }
     }
 
     @Subscribe
     public void onRemoveByPath(RemoveByPathEvent event) {
+        if (event == null) {
+            Log.e(TAG, "Null event");
+            return;
+        }
+        if (repository == null) {
+            Log.e(TAG, "Null repository");
+            return;
+        }
+        if (event.path == null) {
+            Log.e(TAG, "Null video");
+            return;
+        }
 
         if (event.type == Type.SUMMARISED) {
-            Log.i(TAG, "onRemoveByPath");
-            repository.delete(event.path);
+            Log.v(TAG, "onRemoveByPath");
+            try {
+                repository.delete(event.path);
+            } catch (Exception e) {
+                Log.e(TAG, String.format("onRemoveByPath error: \n%s", e.getMessage()));
+            }
         }
     }
